Create Comment Line Arguments
The command line argument in java is the information passed to the program at the time of running the program. It is the argument passed through the console when the program is run. The command line argument is the data that is written right after the program's name at the command line while executing the program.